+++ client/package.json
@@ -19,6 +19,7 @@
         "recoil": "^0.6.1",
         "styled-components": "^5.3.3",
         "styled-reset": "^4.3.4",
+        "three": "^0.137.5",
         "typescript": "^4.5.5",
         "web-vitals": "^2.1.4"
     },
@@ -47,6 +48,7 @@
         ]
     },
     "devDependencies": {
+        "@types/three": "^0.137.0",
         "customize-cra": "^1.0.0",
         "react-app-rewired": "^2.1.11"
     }
