+++ client/src/App.tsx
@@ -1,6 +1,6 @@
-import { Switch } from "react-router-dom";
-import PublicRoute from "@Routes/PublicRoute";
-import PrivateRoute from "@Routes/PrivateRoute";
+import { Routes } from "react-router-dom";
+import PublicRoute from "@Route/PublicRoute";
+import PrivateRoute from "@Route/PrivateRoute";
 import MainPage from "@Pages/Main";
 import LoginPage from "@Pages/Login.jsx";
 import MapPage from "@Pages/Map";
@@ -9,13 +9,13 @@ import TestPage2 from "@Pages/Test2";
 
 const App = () => {
     return (
-        <Switch>
-            <PublicRoute path="/" component={MainPage} exact />
-            <PublicRoute path="/login" component={LoginPage} exact />
-            <PrivateRoute path="/map" component={MapPage} exact />
-            <PrivateRoute path="/map1" component={TestPage} exact />
-            <PrivateRoute path="/map2" component={TestPage2} exact />
-        </Switch>
+        <Routes>
+            <PublicRoute path="/" component={MainPage} />
+            <PublicRoute path="/login" component={LoginPage} />
+            <PrivateRoute path="/map" component={MapPage} />
+            <PrivateRoute path="/map1" component={TestPage} />
+            <PrivateRoute path="/map2" component={TestPage2} />
+        </Routes>
     );
 };
