+++ client/src/Common/Constant/index.ts
@@ -1,0 +1,1 @@
+export const hello = () => "hello";
