+++ client/src/Common/Type/index.ts
@@ -1,0 +1,32 @@
+export interface RouterProps {
+    component: () => JSX.Element;
+    path: string;
+}
+
+interface BasicAtomProps {
+    width?: string;
+    height?: string;
+    borderColor?: string;
+}
+
+export interface BasicButtonProps extends BasicAtomProps {
+    backGroundColor?: string;
+    title?: string;
+    onClick?: () => void;
+}
+
+export interface LoginButtonProps extends BasicButtonProps {
+    color?: string;
+    fontSize?: string;
+    radius?: string;
+}
+
+export interface BasicInputProps extends BasicAtomProps {
+    margin?: string;
+    padding?: string;
+    placeholder?: string;
+    background?: string;
+    type?: string;
+    inputRef?: any;
+    onChange?: () => void;
+}
