--- client/src/Components/Atoms/Button/Login.jsx
@@ -1,8 +1,0 @@
-import { LoginContainer } from "./styles";
-
-const LoginButton = (props) => {
-  const { title } = props;
-  return <LoginContainer {...props}>{title}</LoginContainer>;
-};
-
-export default LoginButton;
