--- client/src/Components/Atoms/Button/index.jsx
@@ -1,8 +1,0 @@
-import { ButtonContainer } from "./styles";
-
-const Button = (props) => {
-  const { title } = props;
-  return <ButtonContainer {...props}>{title}</ButtonContainer>;
-};
-
-export default Button;
