--- client/src/Components/Atoms/Input/index.jsx
@@ -1,8 +1,0 @@
-import Container from "./styles";
-
-const Input = (props) => {
-  const { inputRef } = props;
-  return <Container {...props} ref={inputRef} />;
-};
-
-export default Input;
