--- client/src/Routes/PublicRoute.jsx
@@ -1,14 +1,0 @@
-import { Redirect, Route } from "react-router-dom";
-
-const PublicRoute = ({ component: Component, path, exact }) => {
-  const user = false;
-  return (
-    <Route
-      exact={exact}
-      path={path}
-      render={() => (user ? <Redirect to="/map" /> : <Component />)}
-    />
-  );
-};
-
-export default PublicRoute;
